requirements.txt
beautifulsoup4==4.13.4
newspaper==0.1.0.7
nltk==3.9.1
numpy==2.3.1
pandas==2.3.1
python-dotenv==1.1.1
Requests==2.32.4
scikit_learn==1.7.1
spacy==3.8.7
Unidecode==1.4.0
